requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
ollama==0.1.6
requests==2.31.0
email-validator>=2.0.0
mysql-connector-python==8.2.0
